requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.12

huggingface-hub>=1.4.0
sentence-transformers>=5.2.2
transformers>=5.1.0
tokenizers>=0.22.0
torch>=2.2.0

chromadb>=1.5.0
langchain>=0.3.0
langchain-community>=0.3.0

pypdf2>=3.0.1
pdfplumber>=0.11.0

python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.6.0
rank-bm25>=0.2.2

pytest>=8.0.0
black>=24.0.0